Rename server.py to client.py, add logging and DELETE request handling
